feat: 당첨률 소숫점 둘째 자리 반올림

+++ src/main/java/lotto/controller/LottoController.java
@@ -27,7 +27,7 @@ public class LottoController {
         Result result = new Result(lottoBuyer);
         result.calculateResult(winningNumbers,bonusNumber);
         outputView.printWinningResult(result);
-        float winningRate = result.calculateWinningRate();
+        double winningRate = result.calculateWinningRate();
     }
